Variables
Shouldn't HP be an int? We want integer values of HP. Same for attack. 
Also, character will have much more variables than enemies. We need HP, MP, ATK, body parts, buffs, and more. We should have a specific class for character info.

+++ RPG/src/main/Character.java
@@ -12,8 +12,8 @@ public class Character {
 		head
 	 }
 	
-	float currentHP;
-	float maxHP;
-	float attackStrength;
+	int currentHP;
+	int maxHP;
+	int attackStrength;
 	Vector<BodyParts> bodyParts;
 }
